requirements.txt
flask
pydub
openai
python-dotenv
gunicorn
requests
